multiple product

+++ app/Http/Controllers/WishlistController.php
@@ -45,13 +45,26 @@ class WishlistController extends Controller
 			'name' => $request['name'],
 			'client_id' => $client_id
 		]);
-		ProductWishList::create([
-			'client_id' => $client_id,
-			'product_id' => $request->product_id,
-			'wishlist_id' => $wishlist->id,
-			'price' => $product_price,
-			'margin_price' => $product_price
-		]);
+		if(@$request['multipleProduct']){
+				foreach($request['multipleProduct'] as $value){
+					ProductWishList::create([
+			 			'client_id' => $client_id,
+			 			'product_id' => $value['product_id'],
+			 			'wishlist_id' => $wishlist->id,
+			 			'price' => $value['price'],
+			 			'margin_price' => $value['price']
+				 ]);
+			}
+		}else{
+			ProductWishList::create([
+			 'client_id' => $client_id,
+			 'product_id' => $request->product_id,
+			 'wishlist_id' => $wishlist->id,
+			 'price' => $product_price,
+			 'margin_price' => $product_price
+			]);
+		}
+		
 		 $count = ProductWishList::where('client_id',$client_id)->count();
         \Session::put('wishlistCount',$count);
 		 
@@ -65,16 +78,26 @@ class WishlistController extends Controller
 	public function getUserWishList($product_id){
 		if (\Auth::user()){
 			$client_id = \Auth()->user()->id;
-			$wishlist = Wishlist::with('ProductWishList')->where('client_id',$client_id)->get();
+			if($product_id == 0){
+					$wishlist = Wishlist::where('client_id',$client_id)->get();
+			}else{
+					$wishlist = Wishlist::with('ProductWishList')->where('client_id',$client_id)->get();
+			}
+		
 			$html = '';
 
 			foreach($wishlist as $value){
-				$product_array = $value->ProductWishList->pluck('product_id')->toArray();
-				if(in_array($product_id,$product_array)){
-					$checked = 'checked=checked';
-				}else{
+				if($product_id  == 0){
 					$checked = '';
+				}else{
+					$product_array = $value->ProductWishList->pluck('product_id')->toArray();
+					if(in_array($product_id,$product_array)){
+						$checked = 'checked=checked';
+					}else{
+						$checked = '';
+					}
 				}
+				
 			
 				$html.= '<li><input type="checkbox" name="wishlist[]" id="wishlist" class="wishlist" value="'.$value['id'].'" '.$checked.'/>'.$value['name'].'</li>';
 
@@ -90,31 +113,56 @@ class WishlistController extends Controller
 		$product_id =  $request->product_id;
 		$client_id = \Auth()->user()->id;
 		$wishlist = Wishlist::where('id',$wishlist_id)->first();
-		$matchThese = [
-			'wishlist_id' => $wishlist_id,
-			'product_id' => $product_id,
-			'client_id' => $client_id
-		];
-		$productWishList = ProductWishList::where($matchThese)->first();
-		if($productWishList){
-			$productWishList->delete();
-		}else{
-			$matchThese['price'] = $request['product_price'];
-			if(@$wishlist['margin_type']){
-				if($wishlist['margin_type'] == 'percent'){
-					$calculate = ($request['product_price'] / 100) * $wishlist['margin_value'];
-					$matchThese['margin_price'] = $request['product_price'] + $calculate;
-				}else{
-					$matchThese['margin_price'] = $request['product_price'] + $wishlist['margin_value'];
-				}
+		if(@$request['multipleProduct']){
+				foreach($request['multipleProduct'] as $value){
+					$matchThese = [
+							'wishlist_id' => $wishlist_id,
+							'product_id' => $value['product_id'],
+							'client_id' => $client_id
+				 	];
+				 	$productWishList = ProductWishList::where($matchThese)->first();
+				 	if(!$productWishList){
+				 			$matchThese['price'] = $value['price'];
+							if(@$wishlist['margin_type']){
+						if($wishlist['margin_type'] == 'percent'){
+							$calculate = ($value['price'] / 100) * $wishlist['margin_value'];
+							$matchThese['margin_price'] = $value['price'] + $calculate;
+						}else{
+							$matchThese['margin_price'] = $value['price'] + $wishlist['margin_value'];
+						}
+							}
+							ProductWishList::create($matchThese);
+					}
+				
 			}
-			
+		}else{
+				$matchThese = [
+					'wishlist_id' => $wishlist_id,
+					'product_id' => $product_id,
+					'client_id' => $client_id
+				];
 
-			ProductWishList::create($matchThese);
+				$productWishList = ProductWishList::where($matchThese)->first();
+				
+				if($productWishList){
+					$productWishList->delete();
+				}else{
+					$matchThese['price'] = $request['product_price'];
+					if(@$wishlist['margin_type']){
+						if($wishlist['margin_type'] == 'percent'){
+							$calculate = ($request['product_price'] / 100) * $wishlist['margin_value'];
+							$matchThese['margin_price'] = $request['product_price'] + $calculate;
+						}else{
+							$matchThese['margin_price'] = $request['product_price'] + $wishlist['margin_value'];
+						}
+					}
+					ProductWishList::create($matchThese);
+				}
 		}
-        $count = ProductWishList::where('client_id',$client_id)->count();
-        \Session::put('wishlistCount',$count);
-        return $count;
+		
+     $count = ProductWishList::where('client_id',$client_id)->count();
+     \Session::put('wishlistCount',$count);
+     return $count;
 	}
 
 	public function wishlist(){
